+++ src/retrievals/models/embedding_auto.py
@@ -2,8 +2,6 @@ import copy
 import logging
 import os
 import time
-from abc import ABC, abstractmethod
-from contextlib import nullcontext
 from typing import TYPE_CHECKING, Callable, Dict, List, Literal, Optional, Tuple, Union
 
 import numpy as np
@@ -23,6 +21,7 @@ from transformers import (
     PreTrainedTokenizer,
 )
 
+from .base import Base
 from .pooling import AutoPooling
 from .utils import (
     batch_to_device,
@@ -37,63 +36,7 @@ if TYPE_CHECKING:
 logger = logging.getLogger(__name__)
 
 
-class BaseEmbedding(ABC, torch.nn.Module):
-    def __init__(
-        self,
-        model: Optional[nn.Module] = None,
-        tokenizer: Optional[PreTrainedTokenizer] = None,
-        pooling_method: str = 'cls',
-        **kwargs,
-    ):
-        super().__init__()
-        if isinstance(model, str):
-            assert ValueError("Please use AutoModelForEmbedding.from_pretrained(model_name_or_path)")
-        self.model = model
-        self.tokenizer = tokenizer
-        self.pooling_method = pooling_method
-
-    @abstractmethod
-    def forward(self, *args, **kwargs):
-        """Pytorch forward method."""
-        pass
-
-    @abstractmethod
-    def encode(self, *args, **kwargs):
-        """Encode documents."""
-        pass
-
-    def _encode_from_loader(
-        self,
-        loader: DataLoader,
-        convert_to_numpy: bool = True,
-        device: str = None,
-        normalize_embeddings: bool = False,
-        show_progress_bar: bool = None,
-        **kwargs,
-    ) -> Union[List[torch.Tensor], np.ndarray, torch.Tensor]:
-        device = device or self.device
-        self.model.eval()
-        self.model.to(device)
-
-        all_embeddings = []
-
-        for idx, inputs in enumerate(tqdm(loader, desc="Encoding", disable=not show_progress_bar)):
-            with torch.autocast(device_type=device) if self.use_fp16 else nullcontext():
-                with torch.no_grad():
-                    inputs_on_device = {k: v.to(device) for k, v in inputs.items()}
-                    embeddings = self.forward_from_loader(inputs_on_device)
-                    embeddings = embeddings.detach()
-                    if normalize_embeddings:
-                        embeddings = torch.nn.functional.normalize(embeddings, p=2, dim=1)
-                    all_embeddings.append(embeddings)
-        if convert_to_numpy:
-            all_embeddings = np.concatenate([emb.cpu().numpy() for emb in all_embeddings], axis=0)
-        else:
-            all_embeddings = torch.concat(all_embeddings)
-        return all_embeddings
-
-
-class AutoModelForEmbedding(BaseEmbedding):
+class AutoModelForEmbedding(Base):
     """
     Loads or creates an Embedding model that can be used to map sentences / text.
